+++ lib/editorjs_renderer/blocks/paragraph.rb
@@ -1,21 +1,21 @@
 # frozen_string_literal: true
 
-# lib/editor_rails/blocks/paragraph.rb
-module EditorRails
+# lib/editorjs_renderer/blocks/paragraph.rb
+module EditorjsRenderer
   module Blocks
     # Paragraph block rendering logic.
     # Converts "paragraph" block data into a `<p>` HTML element and plain text string.
     #
     # Expects schema to define at least a "text" property in the data.
     #
-    # @see EditorRails::Blocks::Base
+    # @see EditorjsRenderer::Blocks::Base
     class Paragraph < Base
-      def render
-        "<p>#{ERB::Util.html_escape(data["text"])}</p>"
+      def to_html
+        "<p>#{ERB::Util.html_escape(block_data["text"])}</p>"
       end
 
-      def plain
-        data["text"].to_s
+      def to_plain
+        block_data["text"].to_s
       end
     end
   end