+++ lib/editorjs_renderer/renderers/plain_renderer.rb
@@ -1,13 +1,13 @@
 # frozen_string_literal: true
 
-# lib/editor_rails/renderers/plain_renderer.rb
-module EditorRails
+# lib/editorjs_renderer/renderers/plain_renderer.rb
+module EditorjsRenderer
   module Renderers
     # Renders an array of block objects to plain text output.
     # Used for indexing, previews, or fallback content.
     module PlainRenderer
       def self.render(blocks)
-        blocks.map(&:plain).join("\n")
+        blocks.map(&:to_plain).join("\n")
       end
     end
   end